chore: remove compiled artifacts
